Start extracting fgb code to sub-modules
