+++ src/geoq/mod.rs
@@ -14,3 +14,4 @@ pub mod par;
 pub mod reader;
 pub mod simplify;
 pub mod text;
+pub mod fgb;
